+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/code/operations/GraphNodeOperation.java
@@ -1,7 +1,6 @@
 package nl.ou.refactoring.advice.nodes.code.operations;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -20,6 +19,7 @@ import nl.ou.refactoring.advice.nodes.GraphNodeBase;
 import nl.ou.refactoring.advice.nodes.code.GraphNodeType;
 import nl.ou.refactoring.advice.nodes.code.classes.GraphNodeClassMember;
 import nl.ou.refactoring.advice.nodes.code.operations.expressions.GraphNodeMethodInvocationExpression;
+import nl.ou.refactoring.advice.nodes.code.tokens.GraphNodeIdentifier;
 import nl.ou.refactoring.advice.nodes.workflow.microsteps.GraphNodeMicrostepAddMethod;
 import nl.ou.refactoring.advice.nodes.workflow.microsteps.GraphNodeMicrostepRemoveMethod;
 
@@ -27,39 +27,68 @@ import nl.ou.refactoring.advice.nodes.workflow.microsteps.GraphNodeMicrostepRemo
  * Represents a node in a Refactoring Advice Graph that describes an Operation of a Class that is affected by a refactoring.
  */
 public final class GraphNodeOperation extends GraphNodeClassMember {
-	private final String operationName;
-	private final List<GraphNodeOperationParameter> operationParameters;
+	private final GraphEdgeHas operationNameEdge;
+	private final GraphEdgeHas operationParameterHead;
 	
 	/**
 	 * Initialises a new instance of {@link GraphNodeOperation}.
 	 * @param graph The graph that contains the node.
-	 * @param operationName The name of the operation.
+	 * @param operationName The node that represents the name of the operation.
 	 * @param operationParameters The parameters of the operation. If null, an empty list will be created.
-	 * @throws ArgumentNullException Thrown if graph or operationName are null.
+	 * @throws ArgumentNullException Thrown if graph, operationName or operationParameters are null.
 	 * @throws ArgumentEmptyException Thrown if operationName is empty or contains only white spaces.
 	 */
 	public GraphNodeOperation(
 			Graph graph,
-			String operationName,
+			GraphNodeIdentifier operationName,
 			List<GraphNodeOperationParameter> operationParameters
 	) throws ArgumentNullException, ArgumentEmptyException {
+		ArgumentGuard.requireNotNull(graph, "graph");
+		ArgumentGuard.requireNotNull(operationName, "operationName");
+		ArgumentGuard.requireNotNull(operationParameters, "operationParameters");
 		super(graph);
-		ArgumentGuard.requireNotNullEmptyOrWhiteSpace(operationName, "operationName");
-		this.operationName = operationName;
-		this.operationParameters =
-			operationParameters == null
-				? new ArrayList<GraphNodeOperationParameter>()
-				: operationParameters;
+		this.operationNameEdge =
+			this.graph.getOrAddEdge(
+				this,
+				operationName,
+				(source, destination) -> new GraphEdgeHas(source, destination),
+				GraphEdgeHas.class
+			);
+		
+		if (operationParameters.size() == 0) {
+			this.operationParameterHead = null;
+		} else {
+			var node = operationParameters.get(0);
+			if (node == null) {
+				throw new GraphNodeOperationParameterNullReferenceException(this, 0);
+			}
+			this.operationParameterHead =
+				this.graph.getOrAddEdge(
+					this,
+					operationParameters.get(0),
+					(source, destination) -> new GraphEdgeHas(source, destination),
+					GraphEdgeHas.class
+				);
+			GraphNodeOperationParameter nodeNext;
+			for (var i = 1; i < operationParameters.size(); i++) {
+				nodeNext = operationParameters.get(i);
+				if (nodeNext == null) {
+					throw new GraphNodeOperationParameterNullReferenceException(this, i);
+				}
+				node.hasNext(nodeNext);
+				nodeNext = node;
+			}
+		}
 	}
 	
 	/**
 	 * Initialises a new instance of {@link GraphNodeOperation}.
 	 * @param graph The graph that contains the node.
-	 * @param operationName The name of the operation.
+	 * @param operationName The node that represents the name of the operation.
 	 * @throws ArgumentNullException Thrown if graph or operationName are null.
 	 * @throws ArgumentEmptyException Thrown if operationName is empty or contains only white spaces.
 	 */
-	public GraphNodeOperation(Graph graph, String operationName)
+	public GraphNodeOperation(Graph graph, GraphNodeIdentifier operationName)
 			throws ArgumentNullException, ArgumentEmptyException {
 		this(graph, operationName, new ArrayList<GraphNodeOperationParameter>());
 	}
@@ -88,7 +117,7 @@ public final class GraphNodeOperation extends GraphNodeClassMember {
 	 * @return The name of the Operation affected by a refactoring.
 	 */
 	public String getOperationName() {
-		return this.operationName;
+		return this.operationNameEdge.getDestinationNode().toString();
 	}
 	
 	/**
@@ -127,7 +156,16 @@ public final class GraphNodeOperation extends GraphNodeClassMember {
 	 * @return The Operation Parameters of this Operation.
 	 */
 	public List<GraphNodeOperationParameter> getOperationParameters() {
-		return Collections.unmodifiableList(this.operationParameters);
+		if (this.operationParameterHead == null) {
+			return List.of();
+		}
+		final var operationParameterList = new ArrayList<GraphNodeOperationParameter>();
+		var node = (GraphNodeOperationParameter)this.operationParameterHead.getDestinationNode();
+		while (node != null) {
+			operationParameterList.add(node);
+			node = node.getNext();
+		}
+		return operationParameterList;
 	}
 	
 	/**
@@ -194,8 +232,18 @@ public final class GraphNodeOperation extends GraphNodeClassMember {
 	@Override
 	public GraphNodeBase clone(Graph graph) throws ArgumentNullException {
 		ArgumentGuard.requireNotNull(graph, "graph");
-		// TODO include parameters
-		return new GraphNodeOperation(graph, this.operationName);
+		final var operationNameClone = (GraphNodeIdentifier)this.operationNameEdge.getDestinationNode().clone(graph);
+		final var operationParameterClones =
+			this
+				.getOperationParameters()
+				.stream()
+				.map(node -> (GraphNodeOperationParameter)node.clone(graph))
+				.collect(Collectors.toUnmodifiableList());
+		return new GraphNodeOperation(
+			graph,
+			operationNameClone,
+			operationParameterClones
+		);
 	}
 	
 	@Override
@@ -234,6 +282,6 @@ public final class GraphNodeOperation extends GraphNodeClassMember {
 
 	@Override
 	public String getCaption() {
-		return this.operationName;
+		return this.getOperationName();
 	}
 }