+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/code/operations/expressions/GraphNodeFieldAccess.java
@@ -1,5 +1,7 @@
 package nl.ou.refactoring.advice.nodes.code.operations.expressions;
 
+import nl.ou.refactoring.advice.nodes.code.tokens.GraphNodeIdentifier;
+
 /**
  * A node in a Refactoring Advice Graph that represents a field access expression.
  */
@@ -7,4 +9,9 @@ public interface GraphNodeFieldAccess
 	extends
 		GraphNodeLeftHandSide,
 		GraphNodePrimaryNoNewArrayExpression {
+	/**
+	 * Gets the node that represents the identifier of the field that is accessed.
+	 * @return The node that represents the identifier of the field that is accessed.
+	 */
+	public GraphNodeIdentifier getIdentifier();
 }