+++ nl.ou.refactoring.advice/src/test/java/nl/ou/refactoring/advice/io/plantuml/classDiagrams/GraphPlantUmlClassDiagramWriterTests.java
@@ -23,6 +23,7 @@ import nl.ou.refactoring.advice.nodes.code.GraphNodeType;
 import nl.ou.refactoring.advice.nodes.code.classes.GraphNodeClass;
 import nl.ou.refactoring.advice.nodes.code.classes.GraphNodeClassHasMultipleGeneralisationsException;
 import nl.ou.refactoring.advice.nodes.code.operations.GraphNodeOperation;
+import nl.ou.refactoring.advice.nodes.code.tokens.GraphNodeIdentifier;
 import nl.ou.refactoring.advice.nodes.workflow.RefactoringMayContainOnlyOneStartNodeException;
 import nl.ou.refactoring.advice.nodes.workflow.microsteps.GraphNodeMicrostepAddExpression;
 import nl.ou.refactoring.advice.nodes.workflow.microsteps.GraphNodeMicrostepAddMethod;
@@ -61,7 +62,8 @@ public final class GraphPlantUmlClassDiagramWriterTests {
 		final var classAlpha = new GraphNodeClass(graph, "Alpha");
 		final var attributeAlphaFoo = new GraphNodeAttribute(graph, "foo");
 		final var attributeAlphaBar = new GraphNodeAttribute(graph, "bar");
-		final var operationAlphaAbc = new GraphNodeOperation(graph, "abc");
+		final var abcIdentifier = new GraphNodeIdentifier(graph, "abc");
+		final var operationAlphaAbc = new GraphNodeOperation(graph, abcIdentifier);
 		packageRefactoring.has(classAlpha);
 		classAlpha.has(operationAlphaAbc);
 		classAlpha.has(attributeAlphaFoo);
@@ -71,7 +73,8 @@ public final class GraphPlantUmlClassDiagramWriterTests {
 		operationAlphaAbc.hasReturnType(typeString);
 		final var classBeta = new GraphNodeClass(graph, "Beta");
 		final var attributeBetaMyField = new GraphNodeAttribute(graph, "myField");
-		final var operationBetaAbc2 = new GraphNodeOperation(graph, "abc2");
+		final var abc2Identifier = new GraphNodeIdentifier(graph, "abc2");
+		final var operationBetaAbc2 = new GraphNodeOperation(graph, abc2Identifier);
 		packageRefactoring.has(classBeta);
 		classBeta.has(operationBetaAbc2);
 		classBeta.has(attributeBetaMyField);