+++ nl.ou.refactoring.advice/src/test/java/nl/ou/refactoring/advice/nodes/code/GraphNodeClassTests.java
@@ -15,6 +15,7 @@ import nl.ou.refactoring.advice.Graph;
 import nl.ou.refactoring.advice.nodes.code.classes.GraphNodeClass;
 import nl.ou.refactoring.advice.nodes.code.operations.GraphNodeOperation;
 import nl.ou.refactoring.advice.nodes.code.operations.GraphNodeOperationParameter;
+import nl.ou.refactoring.advice.nodes.code.tokens.GraphNodeIdentifier;
 
 public final class GraphNodeClassTests {
 
@@ -22,10 +23,10 @@ public final class GraphNodeClassTests {
 	@MethodSource("getOperationNodeTestCases")
 	@DisplayName("Should get the requested operation node if present, otherwise return null")
 	public void getOperationNodeTest(
-			GraphNodeClass classNode,
-			String operationName,
-			List<GraphNodeOperationParameter> operationParameters,
-			GraphNodeOperation expectedOperationNode
+		GraphNodeClass classNode,
+		GraphNodeIdentifier operationName,
+		List<GraphNodeOperationParameter> operationParameters,
+		GraphNodeOperation expectedOperationNode
 	) {
 		final var actualOperationNode = classNode.getOperationNode(operationName, operationParameters);
 		assertEquals(expectedOperationNode, actualOperationNode);
@@ -36,7 +37,7 @@ public final class GraphNodeClassTests {
 		
 		final var graph1 = new Graph("Graph test simple operation");
 		final var graph1ClassNode = new GraphNodeClass(graph1, "Class 1");
-		final var graph1OperationNodeName = "Operation 1";
+		final var graph1OperationNodeName = new GraphNodeIdentifier(graph1, "operation1");
 		final var graph1OperationNodeParameters = new ArrayList<GraphNodeOperationParameter>();
 		final var graph1OperationNode = new GraphNodeOperation(graph1, graph1OperationNodeName);
 		graph1ClassNode.has(graph1OperationNode);
@@ -51,7 +52,7 @@ public final class GraphNodeClassTests {
 		
 		final var graph2 = new Graph("Graph test operation with parameters");
 		final var graph2ClassNode = new GraphNodeClass(graph2, "Class 2");
-		final var graph2OperationNodeName = "Operation 2";
+		final var graph2OperationNodeName = new GraphNodeIdentifier(graph2, "operation2");
 		final var graph2OperationNodeParametersInput = new ArrayList<GraphNodeOperationParameter>();
 		graph2OperationNodeParametersInput.add(new GraphNodeOperationParameter(graph2, "test"));
 		final var graph2OperationNodeParametersModel = new ArrayList<GraphNodeOperationParameter>();
@@ -74,7 +75,7 @@ public final class GraphNodeClassTests {
 		
 		final var graph3 = new Graph("Graph test operation not found");
 		final var graph3ClassNode = new GraphNodeClass(graph3, "Class 3");
-		final var graph3OperationNodeName = "Operation 3";
+		final var graph3OperationNodeName = new GraphNodeIdentifier(graph3, "operation3");
 		final var graph3OperationNodeParameters = new ArrayList<GraphNodeOperationParameter>();
 		final GraphNodeOperation graph3OperationNode = null;
 		argumentsList.add(
@@ -88,7 +89,7 @@ public final class GraphNodeClassTests {
 		
 		final var graph4 = new Graph("Graph test operation parameters mismatch");
 		final var graph4ClassNode = new GraphNodeClass(graph4, "Class 4");
-		final var graph4OperationNodeName = "Operation 4";
+		final var graph4OperationNodeName = new GraphNodeIdentifier(graph4, "operation4");
 		final var graph4OperationNodeParametersInput = new ArrayList<GraphNodeOperationParameter>();
 		graph4OperationNodeParametersInput.add(new GraphNodeOperationParameter(graph4, "test"));
 		final var graph4OperationNodeParametersModel = new ArrayList<GraphNodeOperationParameter>();